+++ src/components/drawer/drawer.stories.tsx
@@ -888,7 +888,7 @@ export const SideViewNavigation: Story = () => {
   const [isExpanded, setIsExpanded] = useState(true);
   const [isFilterOpen, setFilterOpen] = useState(false);
   const [isDialogOpen, setIsDialogOpen] = useState(false);
-  const [sortType, setSortType] = useState<"ascending" | "descending" | false>(
+  const [sortType, setSortType] = useState<"ascending" | "descending">(
     "descending"
   );
   const [pickedUpData, setPickedUpData] = useState<dataPropTypes>();