+++ src/locales/en-gb.ts
@@ -151,6 +151,9 @@ const enGB: Locale = {
   pod: {
     undo: () => "Undo",
   },
+  search: {
+    searchButtonText: () => "Search",
+  },
   select: {
     actionButtonText: () => "Add New Item",
     placeholder: () => "Please Select...",
@@ -160,6 +163,14 @@ const enGB: Locale = {
     ariaLabels: {
       close: () => "Close",
     },
+  },
+  sort: {
+    accessibleName: (sortContent, sortType) =>
+      `Sort all ${sortContent || "contents"}${
+        sortType
+          ? ` in an ${sortType} order.`
+          : " in an ascending or descending order."
+      }`,
   },
   splitButton: {
     ariaLabel: () => "Show more",